parse logs to jsonl only
